api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.3.7
qdrant-client==1.6.0
pydantic==2.5.2
